# fix: Don't instantiate the scalar composition code quadratically for dictionaries